[weather.multi] 0.0.3
