+++ weather.multi/changelog.txt
@@ -1,2 +1,8 @@
+v0.0.3
+- fix maps
+
+v0.0.2
+- refactor code
+
 v0.0.1
 - initial release
